+++ chaoscenter/graphql/definitions/shared/chaos_experiment.graphqls
@@ -702,18 +702,7 @@ type GetExperimentStatsResponse {
 }
 
 extend type Query {
-  """
-  Returns experiment run based on experiment run ID
-  """
-  getExperimentRun(projectID: ID!, experimentRunID: String!): ExperimentRun!
 
-  """
-  Returns the list of experiment run based on various filter parameters
-  """
-  listExperimentRun(
-    projectID: ID!
-    request: ListExperimentRunRequest!
-  ): ListExperimentRunResponse!
 
   """
   Returns the experiment based on experiment ID
@@ -728,10 +717,6 @@ extend type Query {
     request: ListExperimentRequest!
   ): ListExperimentResponse!
 
-  """
-  Query to get experiment run stats
-  """
-  getExperimentRunStats(projectID: ID!): GetExperimentRunStatsResponse!
   """
   Query to get experiment stats
   """
@@ -756,14 +741,6 @@ extend type Mutation {
   ): String!
 
   """
-  Run the chaos experiment (used by frontend)
-  """
-  runChaosExperiment(
-    experimentID: String!
-    projectID: ID!
-  ): RunChaosExperimentResponse!
-
-  """
   Updates the experiment
   """
   updateChaosExperiment(
@@ -779,10 +756,4 @@ extend type Mutation {
     experimentRunID: String
     projectID: ID!
   ): Boolean!
-
-  """
-  Creates a new experiment run and sends it to subscriber
-  """
-  # authorized directive not required
-  chaosExperimentRun(request: ExperimentRunRequest!): String!
 }
