+++ chaoscenter/graphql/server/graph/chaos_experiment.resolvers.go
@@ -46,7 +46,7 @@ func (r *mutationResolver) CreateChaosExperiment(ctx context.Context, request mo
 
 		if experiment.CronSyntax != "" {
 
-			if err = r.chaosExperimentHandler.RunCronExperiment(ctx, projectID, experiment, data_store.Store); err != nil {
+			if err = r.chaosExperimentRunHandler.RunCronExperiment(ctx, projectID, experiment, data_store.Store); err != nil {
 				logrus.WithFields(logFields).Error(err)
 				return nil, err
 			}
@@ -54,7 +54,7 @@ func (r *mutationResolver) CreateChaosExperiment(ctx context.Context, request mo
 			return uiResponse, nil
 		}
 
-		_, err = r.chaosExperimentHandler.RunChaosWorkFlow(ctx, projectID, experiment, data_store.Store)
+		_, err = r.chaosExperimentRunHandler.RunChaosWorkFlow(ctx, projectID, experiment, data_store.Store)
 		if err != nil {
 			logrus.WithFields(logFields).Error(err)
 			return nil, err
@@ -87,41 +87,6 @@ func (r *mutationResolver) SaveChaosExperiment(ctx context.Context, request mode
 	}
 
 	return uiResponse, nil
-}
-
-func (r *mutationResolver) RunChaosExperiment(ctx context.Context, experimentID string, projectID string) (*model.RunChaosExperimentResponse, error) {
-	logFields := logrus.Fields{
-		"projectId":         projectID,
-		"chaosExperimentId": experimentID,
-	}
-
-	logrus.WithFields(logFields).Info("request received to run chaos experiment")
-	err := authorization.ValidateRole(ctx, projectID,
-		authorization.MutationRbacRules[authorization.CreateChaosWorkFlow],
-		model.InvitationAccepted.String())
-	if err != nil {
-		return nil, err
-	}
-
-	query := bson.D{
-		{"experiment_id", experimentID},
-		{"is_removed", false},
-	}
-
-	experiment, err := r.chaosExperimentHandler.GetDBExperiment(query)
-	if err != nil {
-		return nil, errors.New("could not get experiment run, error: " + err.Error())
-	}
-
-	var uiResponse *model.RunChaosExperimentResponse
-
-	uiResponse, err = r.chaosExperimentHandler.RunChaosWorkFlow(ctx, projectID, experiment, data_store.Store)
-	if err != nil {
-		logrus.WithFields(logFields).Error(err)
-		return nil, err
-	}
-
-	return &model.RunChaosExperimentResponse{NotifyID: uiResponse.NotifyID}, err
 }
 
 func (r *mutationResolver) UpdateChaosExperiment(ctx context.Context, request *model.ChaosExperimentRequest, projectID string) (*model.ChaosExperimentResponse, error) {
@@ -170,53 +135,6 @@ func (r *mutationResolver) DeleteChaosExperiment(ctx context.Context, experiment
 	return uiResponse, err
 }
 
-func (r *mutationResolver) ChaosExperimentRun(ctx context.Context, request model.ExperimentRunRequest) (string, error) {
-	return r.chaosExperimentHandler.ChaosExperimentRunEvent(request)
-}
-
-func (r *queryResolver) GetExperimentRun(ctx context.Context, projectID string, experimentRunID string) (*model.ExperimentRun, error) {
-	logFields := logrus.Fields{
-		"projectId":            projectID,
-		"chaosExperimentRunId": experimentRunID,
-	}
-	logrus.WithFields(logFields).Info("request received to fetch chaos experiment run")
-	err := authorization.ValidateRole(ctx, projectID,
-		authorization.MutationRbacRules[authorization.ListWorkflowRuns],
-		model.InvitationAccepted.String())
-	if err != nil {
-		return nil, err
-	}
-
-	expRunResponse, err := r.chaosExperimentHandler.GetExperimentRun(ctx, projectID, experimentRunID)
-	if err != nil {
-		logrus.WithFields(logFields).Error(err)
-		return nil, err
-	}
-	return expRunResponse, err
-}
-
-func (r *queryResolver) ListExperimentRun(ctx context.Context, projectID string, request model.ListExperimentRunRequest) (*model.ListExperimentRunResponse, error) {
-	logFields := logrus.Fields{
-		"projectId":             projectID,
-		"chaosExperimentIds":    request.ExperimentIDs,
-		"chaosExperimentRunIds": request.ExperimentRunIDs,
-	}
-	logrus.WithFields(logFields).Info("request received to list chaos experiment run")
-
-	err := authorization.ValidateRole(ctx, projectID,
-		authorization.MutationRbacRules[authorization.ListWorkflowRuns],
-		model.InvitationAccepted.String())
-	if err != nil {
-		return nil, err
-	}
-	uiResponse, err := r.chaosExperimentHandler.ListExperimentRun(projectID, request)
-	if err != nil {
-		logrus.WithFields(logFields).Error(err)
-		return nil, err
-	}
-	return uiResponse, err
-}
-
 func (r *queryResolver) GetExperiment(ctx context.Context, projectID string, experimentID string) (*model.GetExperimentResponse, error) {
 	logFields := logrus.Fields{
 		"projectId":         projectID,
@@ -252,26 +170,6 @@ func (r *queryResolver) ListExperiment(ctx context.Context, projectID string, re
 		return nil, err
 	}
 	uiResponse, err := r.chaosExperimentHandler.ListExperiment(projectID, request)
-	if err != nil {
-		logrus.WithFields(logFields).Error(err)
-		return nil, err
-	}
-	return uiResponse, err
-}
-
-func (r *queryResolver) GetExperimentRunStats(ctx context.Context, projectID string) (*model.GetExperimentRunStatsResponse, error) {
-	logFields := logrus.Fields{
-		"projectId": projectID,
-	}
-	logrus.WithFields(logFields).Info("request received to get chaos experiment run stats")
-	err := authorization.ValidateRole(ctx, projectID,
-		authorization.MutationRbacRules[authorization.ListWorkflowRuns],
-		model.InvitationAccepted.String())
-	if err != nil {
-		return nil, err
-	}
-
-	uiResponse, err := r.chaosExperimentHandler.GetExperimentRunStats(ctx, projectID)
 	if err != nil {
 		logrus.WithFields(logFields).Error(err)
 		return nil, err
